Move the part of utils for all model
